[onechat] move built-in tools registration
