+++ x-pack/platform/plugins/shared/onechat/server/services/tools/builtin/builtin_registry.test.ts
