+++ x-pack/platform/plugins/shared/onechat/server/services/tools/builtin/retrieval/execute_esql.ts
