+++ x-pack/platform/plugins/shared/onechat/server/services/tools/builtin/retrieval/generate_esql.ts
