+++ x-pack/platform/plugins/shared/onechat/server/services/tools/builtin/retrieval/get_document_by_id.ts
