+++ x-pack/platform/plugins/shared/onechat/server/services/tools/builtin/retrieval/get_index_mapping.ts
