+++ x-pack/platform/plugins/shared/onechat/server/services/tools/builtin/retrieval/index_explorer.ts
