+++ x-pack/platform/plugins/shared/onechat/server/services/tools/builtin/retrieval/list_indices.ts
