+++ x-pack/platform/plugins/shared/onechat/server/services/tools/builtin/retrieval/nl_search.ts
